+++ README.md
@@ -15,7 +15,7 @@ Think is an educational programming language designed to teach computational thi
 ```bash
 # Clone the repository
 git clone https://github.com/yourusername/think.git
-cd thinkpy
+cd think
 
 # Install the package
 pip install -e .
