+++ think/interpreter.py
