+++ think/jupyter_magic.py
